refactor: Move populate code to own module
Mimesis fails to import if python2 is used to run the app. Since the
database code and populate code were in the same module this breaks
the app.

Moving the populate functions to its own module, catalog/populate.py
fixes this. An assert Python 3 is also added to the new module.

TODO: Documentation will need to be updated
